Added Edit Shop Details Dialog
Found the check to see if the shop is the users not working. Fixed this issue. Modified 'Contact Owner' button to allow editing shop details when logged in. Dialog now appears to edit shop title and description. Requires an end point for the title and description still. I think I can change the file structure of this code somewhat. I would appreciate advice on this topic.